+++ lib/src/key_pair/key_pair.dart
@@ -3,8 +3,14 @@ import 'dart:typed_data';
 import '../types.dart';
 
 abstract interface class KeyPair {
+  Uint8List get privateKey;
+
+  List<SignatureScheme> get supportedSignatureSchemes;
+
   Future<Uint8List> getPublicKey();
+
   Future<KeyType> getKeyType();
+
   Future<String> getKeyId();
 
   Future<Uint8List> sign(