+++ lib/src/wallet/aws_kms_wallet.dart
@@ -2,7 +2,7 @@ import 'dart:typed_data';
 
 import 'package:aws_kms_api/kms-2014-11-01.dart' as kms;
 
-import 'package:affinidi_ssi/src/key_pair/aws_kms_key_pair.dart';
+import 'package:ssi/src/key_pair/aws_kms_key_pair.dart';
 
 import 'wallet.dart';
 import '../types.dart';