+++ test/aws_kms_test.dart
@@ -3,7 +3,7 @@ import 'dart:typed_data';
 import 'package:aws_kms_api/kms-2014-11-01.dart' as kms;
 import 'package:shared_aws_api/shared.dart';
 
-import 'package:affinidi_ssi/affinidi_ssi.dart';
+import 'package:ssi/ssi.dart';
 
 import 'package:test/test.dart';