+++ CHANGELOG.md
@@ -3,6 +3,97 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 2025-04-02
+
+### Changes
+
+---
+
+Packages with breaking changes:
+
+ - There are no breaking changes in this release.
+
+Packages with other changes:
+
+ - [`ssi` - `v1.0.0-dev.10`](#ssi---v100-dev10)
+
+---
+
+#### `ssi` - `v1.0.0-dev.10`
+
+ - **FEAT**: DID handling refactoring (#11).
+
+## 1.0.0-dev.10
+
+ - **FEAT**: DID handling refactoring (#11).
+
+
+## 2025-04-02
+
+### Changes
+
+---
+
+Packages with breaking changes:
+
+ - There are no breaking changes in this release.
+
+Packages with other changes:
+
+ - [`ssi` - `v1.0.0-dev.9`](#ssi---v100-dev9)
+
+---
+
+#### `ssi` - `v1.0.0-dev.9`
+
+ - **FIX**: checks ci reference (#18).
+ - **FIX**: linting and  packaging issues (#5).
+ - **FIX**: implement did.
+ - **FIX**: signature scheme (#2).
+ - **FIX**: format.
+ - **FEAT**: name package & setup ci (#14).
+ - **FEAT**: VC interfaces.
+ - **FEAT**: add base resolution and did web (#8).
+ - **FEAT**: add edward curve add did peer (#7).
+
+## 1.0.0-dev.9
+
+ - **FIX**: checks ci reference (#18).
+ - **FIX**: linting and  packaging issues (#5).
+ - **FIX**: implement did.
+ - **FIX**: signature scheme (#2).
+ - **FIX**: format.
+ - **FEAT**: name package & setup ci (#14).
+ - **FEAT**: VC interfaces.
+ - **FEAT**: add base resolution and did web (#8).
+ - **FEAT**: add edward curve add did peer (#7).
+
+
+## 2025-04-01
+
+### Changes
+
+---
+
+Packages with breaking changes:
+
+ - There are no breaking changes in this release.
+
+Packages with other changes:
+
+ - [`ssi` - `v1.0.0-dev.8`](#ssi---v100-dev8)
+
+---
+
+#### `ssi` - `v1.0.0-dev.8`
+
+ - **FEAT**: VC interfaces.
+
+## 1.0.0-dev.8
+
+ - **FEAT**: VC interfaces.
+
+
 ## 2025-03-28
 
 ### Changes
@@ -15,11 +106,11 @@ Packages with breaking changes:
 
 Packages with other changes:
 
- - [`affinidi_ssi` - `v1.0.0-dev.7`](#affinidi_ssi---v100-dev7)
+ - [`ssi` - `v1.0.0-dev.7`](#ssi---v100-dev7)
 
 ---
 
-#### `affinidi_ssi` - `v1.0.0-dev.7`
+#### `ssi` - `v1.0.0-dev.7`
 
  - **FEAT**: add base resolution and did web (#8).
 
@@ -40,11 +131,11 @@ Packages with breaking changes:
 
 Packages with other changes:
 
- - [`affinidi_ssi` - `v1.0.0-dev.6`](#affinidi_ssi---v100-dev6)
+ - [`ssi` - `v1.0.0-dev.6`](#ssi---v100-dev6)
 
 ---
 
-#### `affinidi_ssi` - `v1.0.0-dev.6`
+#### `ssi` - `v1.0.0-dev.6`
 
  - **FEAT**: add edward curve add did peer (#7).
 
@@ -65,11 +156,11 @@ Packages with breaking changes:
 
 Packages with other changes:
 
- - [`affinidi_ssi` - `v1.0.0-dev.5`](#affinidi_ssi---v100-dev5)
+ - [`ssi` - `v1.0.0-dev.5`](#ssi---v100-dev5)
 
 ---
 
-#### `affinidi_ssi` - `v1.0.0-dev.5`
+#### `ssi` - `v1.0.0-dev.5`
 
  - **FIX**: linting and  packaging issues (#5).
  - **FIX**: implement did.
@@ -96,11 +187,11 @@ Packages with breaking changes:
 
 Packages with other changes:
 
- - [`affinidi_ssi` - `v1.0.0-dev.4`](#affinidi_ssi---v100-dev4)
+ - [`ssi` - `v1.0.0-dev.4`](#ssi---v100-dev4)
 
 ---
 
-#### `affinidi_ssi` - `v1.0.0-dev.4`
+#### `ssi` - `v1.0.0-dev.4`
 
  - **FIX**: format.